♻️ Rename the modules and divide the tests by subcommand lines.
